backend/zip_task/requirements.txt
google-cloud-pubsub
google-cloud-firestore
google-cloud-storage
retry